requirements.txt
Django==6.0.2
djangorestframework==3.15.2
django-cors-headers==4.6.0
Pillow==11.1.0
mistralai>=1.0.0
